soc: microchip: add generic service driver
commit 51b0ed2098daf7579d67b6b0aa218f2794d3c87a from
https://github.com/linux4microchip/linux.git linux-6.1-mchp

Add support for the "generic" services on the PolarFire SoC.

The PolarFire SoC system controller is communicated with via a mailbox.
Several of the functions provided by the system controller do not belong
in a specific subsystem, such as reading the fpga device certificate.
These "generic" services all follow the same format:
  - a command + optional payload sent to the sys controller
  - a status + a payload returned to Linux

Add a driver to implement these services and export an interface
as a char device.

Reviewed-by: Daire McNamara <daire.mcnamara@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>