RISC-V: set memblock allocation direction to bottom-up
commit 0fe79ab7e7b2b2fcad31f261b3c7cbef4c6e7b8b from
https://github.com/linux4microchip/linux.git linux-6.1-mchp+fpga

RISC-V currently allocates memblocks in a top-down manner.
Unfortunately, the first allocations occur in paging_init(), *before*
we have scanned for reserved memory.

Accordingly, if the topmost memblock is destined to be reserved memory,
we may erroneously allocate from it. At best, this means a memory
reservation will later fail, and at worst will crash the kernel (or
clobber some other application).

To avoid this, begin to allocate from the lowest memblock up.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>