RISC-V: configs: convert generic-service to a module for mpfs
commit a9dd9dca6d639eb021a86af12b210e3c2c4e138a from
https://github.com/linux4microchip/linux.git linux-6.1-mchp

Ivan complained about the warning emitted by the generic service driver
as being non-ideal for our default image. To appease him, and general
optics, make the driver a module that must be specifically loaded to
use the system service examples, where the meaning of the warning can
be explained more wordily.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>