mailbox: mpfs: fix an incorrect mask width
commit b80f1951b10471db959253a8ab5ed5f5ea2a5a0b from
https://github.com/linux4microchip/linux.git linux-6.1-mchp+fpga

The system controller registers on PolarFire SoC are 32 bits wide, so
16 + 16 as the first input to GENMASK_ULL() gives a 33 bit wide mask.
It probably should have been immediately obvious when it was pointed
out during review that the width required using GENMASK_ULL() - but I
scarcely knew what I was doing at the time and missed it.
The mistake ends up being moot as it is a mask after all, but it is
incorrect and should be fixed.

No functional change intended.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>