riscv: dts: microchip: add ccc clock controller to mpfs
commit 4b7c82440e983cfea07e95b3c9e4bad4ed364c3a from
https://github.com/linux4microchip/linux.git linux-6.1-mchp

The CCCs (Clock Conditioning Circuitry) are clock controllers providing
clocks for the FPGA fabric and peripherals such as PCIe.
Which CCC is used is determined during the creation of an FPGA bitstream
but can be fixed to a particular CCC using design constraints.

Add the CCC used by PCIe on the Icicle kit to the devicetree.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>