soc: sifive: use an arch, not device initcall
commit 72768166708ed879ce7a215576c6e6be3dbb0e1c from
https://github.com/linux4microchip/linux.git linux-6.1-mchp

The current device_initcall() triggers the probing of the SiFive
Composable Cache driver too late in boot, as the PCIe subsystem will
already have probed by this point - triggering a WARN() from core RISC-V
dma handling code, as we will have attempted to configure a device as
non-coherent before the non-coherent DMA handling has been loaded.

Convert the device_initcall() to an arch_initcall() to ensure that the
driver is loaded in time.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>