RISC-V: configs: enable IO_STRICT_DEVMEM for mpfs
commit e0f04c725d2f6f6b0401ad78a148f810d75fe3bf from
https://github.com/linux4microchip/linux.git linux-6.1-mchp

If we're going to be pointing people at the /dev/mem tools for use with
some of the apps demos, we should ensure that (by default at least)
noone is using /dev/mem to fiddle with peripherals that already have
drivers bound.
Enable IO_STRICT_DEVMEM to block those types of access.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>