fpga: auto-update: refuse auto-update if no golden image
commit a355c32b5804eeaff07e5c269b50def56683f065 from
https://github.com/linux4microchip/linux.git linux-6.1-mchp

If there is no golden image, or more accurately if the "pointer" to the
golden image is u32_max, the system controller will lock up if Auto
Update is requested.

Ensure that there is a Golden Image, or that the system controller has
been sufficiently mislead, before initiating the service.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>