Add usersapce mappable DMA buffer device driver
u-dma-buf is a Linux device driver that allocates contiguous memory
blocks in the kernel space as DMA buffers and makes them available from
the user space.

The predecessor of u-dma-buf is udmabuf. The kernel module name has been
changed from "udmabuf" to "u-dma-buf". The purpose of this is to avoid
duplicate names because another kernel module with the same name as
"udmabuf" has been added since Linux Kernel 5.x.

Author: Ichiro Kawazome <ichiro_k@ca2.so-net.ne.jp>
Signed-off-by: Saravanan.K.S <saravanan.kadambathursubramaniyam@windriver.com>